Move transpile API into `qiskit.transpiler`
